Created a `linebreaks_iter()` variant of `linebreaks` for increased flexibility
`linebreaks` has a very simple interface - it takes a `&str` and returns an iterator returning break opportunities tagged by positions.  This change introduces a `linebreaks_iter()` variant that instead of taking a `&str`:
- Takes an iterator passing in `char` and arbitary indexes that do not necessarily have to be `usize`
- A `final_idx` parameter containing the final index.

The original `linebreaks()` function can now easily be implemented in terms of `linebreaks_iter()`:
````
pub fn linebreaks(s: &str) -> impl Iterator<Item = (usize, BreakOpportunity)> + Clone + '_ {
    linebreaks_iter(s.char_indices(), s.len())
}
````

This allows the core algorithm to be decoupled from the representation of the string, allowing text to be passed in alternative formats (e.g. - UTF-16 or UTF-8) or possibily a complex data structure where everything is not conveniently a singular string.